Facilitate longer running co-runners

+++ xRTOS_MMU_SEMAPHORE/main.c
@@ -134,18 +134,18 @@ void core0(void* pParam) {
 	int total = 1000;
 	int step = 0;
 	int dir = 1;
-    unsigned int cycles;
+    u64 cycles;
 #ifdef PMU_EVENT_CORE0_1
-    unsigned int event1;
+    u64 event1;
 #endif
 #ifdef PMU_EVENT_CORE0_2
-    unsigned int event2;
+    u64 event2;
 #endif
 #ifdef PMU_EVENT_CORE0_3
-    unsigned int event3;
+    u64 event3;
 #endif
 #ifdef PMU_EVENT_CORE0_4
-    unsigned int event4;
+    u64 event4;
 #endif
     unsigned int iter=1;
     unsigned int offset=0;
@@ -184,7 +184,7 @@ void core0(void* pParam) {
 		count = i*1000000;
 		cycles = report_cycles_countdown(count);
 		cpc = (float) cycles / count;
-		log_info(0, buf, "Countdown was %lu. Cycles spent=%lu\n\r",
+		log_info(0, buf, "Countdown was %lu. Cycles spent=%llu\n\r",
 				 count, cycles);
 		log_info(0, buf, "Measured cycles per count was %.9f\n\r", cpc);
 	}
@@ -279,7 +279,7 @@ void core0(void* pParam) {
 
 #ifdef SCREEN_ENABLE
 		sprintf(&buf[0],
-				"Core 0 Load: %3i%% Task count: %2i Cycle count: %12u",
+				"Core 0 Load: %3i%% Task count: %2i Cycle count: %llu",
 				xLoadPercentCPU(),
 				xTaskGetNumberOfTasks(),
 				cycles);
@@ -287,7 +287,7 @@ void core0(void* pParam) {
 #endif
 
 		log_info(corenum, buf,
-				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 0 cycle_count: %12u iteration: %u offset: %d\n\r",
+				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 0 cycle_count: %llu iteration: %u offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, BENCH_STRING_CORE0, NR_OF_CORES,
 				 cycles, iter, offset);
 
@@ -298,25 +298,25 @@ void core0(void* pParam) {
 		 * zero(!). Beats me. */
 		log_info(corenum, buf, "%#02x\n\r", 0x03);
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE0_1, event1, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE0_2
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE0_2, event2, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE0_3
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE0_3, event3, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE0_4
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE0_4, event4, iter, offset);
 #endif
@@ -333,18 +333,18 @@ void core1(void* pParam) {
 	int total = 1000;
 	int step = 0;
 	int dir = 1;
-    unsigned int cycles;
+    u64 cycles;
 #ifdef PMU_EVENT_CORE1_1
-    unsigned int event1;
+    u64 event1;
 #endif
 #ifdef PMU_EVENT_CORE1_2
-    unsigned int event2;
+    u64 event2;
 #endif
 #ifdef PMU_EVENT_CORE1_3
-    unsigned int event3;
+    u64 event3;
 #endif
 #ifdef PMU_EVENT_CORE1_4
-    unsigned int event4;
+    u64 event4;
 #endif
     unsigned int iter=1;
     unsigned int offset=0;
@@ -458,7 +458,7 @@ void core1(void* pParam) {
 
 #ifdef SCREEN_ENABLE
 		sprintf(&buf[0],
-				"Core 1 Load: %3i%% Task count: %2i Cycle count: %12u",
+				"Core 1 Load: %3i%% Task count: %2i Cycle count: %llu",
 				xLoadPercentCPU(),
 				xTaskGetNumberOfTasks(),
 				cycles);
@@ -466,31 +466,31 @@ void core1(void* pParam) {
 #endif
 
 		log_info(corenum, buf,
-				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 1 cycle_count: %12u iteration: %u offset: %d\n\r",
+				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 1 cycle_count: %llu iteration: %u offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, BENCH_STRING_CORE1, NR_OF_CORES,
 				 cycles, iter, offset);
 
 #ifdef PMU_EVENT_CORE1_1
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE1_1, event1, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE1_2
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE1_2, event2, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE1_3
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE1_3, event3, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE1_4
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE1_4, event4, iter, offset);
 #endif
@@ -507,18 +507,18 @@ void core2(void* pParam) {
 	int total = 1000;
 	int step = 0;
 	int dir = 1;
-    unsigned int cycles;
+    u64 cycles;
 #ifdef PMU_EVENT_CORE2_1
-    unsigned int event1;
+    u64 event1;
 #endif
 #ifdef PMU_EVENT_CORE2_2
-    unsigned int event2;
+    u64 event2;
 #endif
 #ifdef PMU_EVENT_CORE2_3
-    unsigned int event3;
+    u64 event3;
 #endif
 #ifdef PMU_EVENT_CORE2_4
-    unsigned int event4;
+    u64 event4;
 #endif
     unsigned int iter=1;
     unsigned int offset=0;
@@ -632,7 +632,7 @@ void core2(void* pParam) {
 
 #ifdef SCREEN_ENABLE
 		sprintf(&buf[0],
-				"Core 2 Load: %3i%% Task count: %2i Cycle count: %12u",
+				"Core 2 Load: %3i%% Task count: %2i Cycle count: %llu",
 				xLoadPercentCPU(),
 				xTaskGetNumberOfTasks(),
 				cycles);
@@ -640,31 +640,31 @@ void core2(void* pParam) {
 #endif
 
 		log_info(corenum, buf,
-				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 2 cycle_count: %12u iteration: %u offset: %d\n\r",
+				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 2 cycle_count: %llu iteration: %u offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, BENCH_STRING_CORE2, NR_OF_CORES,
 				 cycles, iter, offset);
 
 #ifdef PMU_EVENT_CORE2_1
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE2_1, event1, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE2_2
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE2_2, event2, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE2_3
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE2_3, event3, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE2_4
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE2_4, event4, iter, offset);
 #endif
@@ -681,18 +681,18 @@ void core3(void* pParam) {
 	int total = 1000;
 	int step = 0;
 	int dir = 1;
-    unsigned int cycles;
+    u64 cycles;
 #ifdef PMU_EVENT_CORE3_1
-    unsigned int event1;
+    u64 event1;
 #endif
 #ifdef PMU_EVENT_CORE3_2
-    unsigned int event2;
+    u64 event2;
 #endif
 #ifdef PMU_EVENT_CORE3_3
-    unsigned int event3;
+    u64 event3;
 #endif
 #ifdef PMU_EVENT_CORE3_4
-    unsigned int event4;
+    u64 event4;
 #endif
     unsigned int iter=1;
     unsigned int offset=0;
@@ -806,7 +806,7 @@ void core3(void* pParam) {
 
 #ifdef SCREEN_ENABLE
 		sprintf(&buf[0],
-				"Core 3 Load: %3i%% Task count: %2i Cycle count: %12u",
+				"Core 3 Load: %3i%% Task count: %2i Cycle count: %llu",
 				xLoadPercentCPU(),
 				xTaskGetNumberOfTasks(),
 				cycles);
@@ -814,31 +814,31 @@ void core3(void* pParam) {
 #endif
 
 		log_info(corenum, buf,
-				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 3 cycle_count: %12u iteration: %u offset: %d\n\r",
+				 "CYCLECOUNT label: %s %s %s %s cores: %d core: 3 cycle_count: %llu iteration: %u offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, BENCH_STRING_CORE3, NR_OF_CORES,
 				 cycles, iter, offset);
 
 #ifdef PMU_EVENT_CORE3_1
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 1 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE3_1, event1, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE3_2
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 2 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE3_2, event2, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE3_3
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 3 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE3_3, event3, iter, offset);
 #endif
 #ifdef PMU_EVENT_CORE3_4
 		log_info(corenum, buf,
-				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %d iteration: %d offset: %d\n\r",
+				 "EVENTCOUNT label: %s %s %s cores: %d core: %d pmu: 4 event_number: %#02x event_count: %llu iteration: %d offset: %d\n\r",
 				 EXP_LABEL, CONFIG_SERIES_STRING, CONFIG_BENCH_STRING, NR_OF_CORES,
 				 corenum, PMU_EVENT_CORE3_4, event4, iter, offset);
 #endif
